requirements.txt
tk
Pillow
numpy